Remove jQuery from embed v3 yepnope
Instead, we'll mention it as a required dependency, but won't try to pull it in ourselves.

Trying to fix an issue where the people working on Balclis' WordPress web site had issues that they think are somehow due to us loading jQuery. Read the chat log: https://auctionet.slack.com/archives/CFQE6HTV2/p1549285944002900

I tried to confirm the issue, but it seemed that would be a lot of work for them to reproduce, and their developer seemed confident about the issue.

My guess is that maybe yepnope loaded our jQuery at a bad time (note the now-removed comment suggesting we didn't handle later-loaded jQuery well), possibly replacing the site's newer jQuery (jQuery().jquery returns "3.3.1") with our old version (1.11.0).

One possible fix would have been to move to a newer jQuery, but they can still conflict, so it would at best fix the current issue for now.

Another idea I considered was to replace jQuery with vanilla JS. I spiked that a bit. Some things are very easy to replace and others (sibling selectors, Ajax) is a little messier. So let's save that until we truly need it.

I'm leaving the Underscore JS and Logan JS dependencies as-is until they become a problem.

I'm changing the v3 in place rather than introduce a v4 because to my knowledge and Tomas's, only Balclis use v3 currently, and they've copied the code rather than hotlink it, so this shouldn't break anything.

This means we're going to have to try to keep our embed code running on all jQuery versions that are in use.